DX-1064: Rename exec_env.rb to context.rb
