+++ .docker/entrypoint/lib/context.rb
