+++ .docker/entrypoint/spec/jekyll_builder_spec.rb
@@ -35,8 +35,8 @@ describe JekyllBuilder do
       site_dir = File.join(data_dir, '_site')
 
       before(:all) do
-        exec_env = ExecEnv.new('development', __dir__, data_dir)
-        jekyll_config_provider = JekyllConfigProvider.new(exec_env, :error)
+        context = ExecEnv.new('development', __dir__, data_dir)
+        jekyll_config_provider = JekyllConfigProvider.new(context, :error)
         jekyll_config = jekyll_config_provider.provide('build')
         jekyll_builder = JekyllBuilder.new(jekyll_config, :error)
         jekyll_builder.execute
